Split document models and tests into separate files
