+++ h/models/document/__init__.py
@@ -1,0 +1,8 @@
+from h.models.document._document import (
+    Document,
+    merge_documents,
+    update_document_metadata,
+)
+from h.models.document._exceptions import ConcurrentUpdateError
+from h.models.document._meta import DocumentMeta, create_or_update_document_meta
+from h.models.document._uri import DocumentURI, create_or_update_document_uri
